añadidos permisos de categorias

+++ app/Http/Controllers/CategoriaController.php
@@ -3,27 +3,33 @@
 namespace App\Http\Controllers;
 
 use App\Models\Categoria;
-use App\Models\Producto;
 use App\Models\Registro;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\File;
 use Illuminate\Validation\Rule;
 
 class CategoriaController extends Controller
 {
     public function index(Request $request)
     {
+        if (request()->user('admin')->cannot('viewAny', Categoria::class)) {
+            return view('admin.categorias.index'); // salir sin enviar datos
+        }
+
         $categorias = Categoria::all();
         return view('admin.categorias.index', compact('categorias'));
     }
 
     public function create()
     {
+        if (request()->user('admin')->cannot('create', Categoria::class)) { abort(403); }
+
         return view('admin.categorias.create');
     }
 
     public function store(Request $request)
     {
+        if (request()->user('admin')->cannot('create', Categoria::class)) { abort(403); }
+
         $request->validate([
             'nombre_categoria'      => ['required', 'max:250', 'unique:categorias'],
             'descripcion_categoria' => [],
@@ -43,11 +49,15 @@ class CategoriaController extends Controller
 
     public function edit(Categoria $categoria)
     {
+        if (request()->user('admin')->cannot('update', Categoria::class)) { abort(403); }
+
         return view('admin.categorias.edit', compact('categoria'));
     }
 
     public function update(Request $request, Categoria $categoria)
     {
+        if (request()->user('admin')->cannot('update', Categoria::class)) { abort(403); }
+
         $request->validate([
             'nombre_categoria' => ['required', 'max:250', Rule::unique('categorias')->ignore($categoria->id)],
             'descripcion_categoria' => [],
@@ -67,6 +77,8 @@ class CategoriaController extends Controller
 
     public function destroy(Categoria $categoria)
     {
+        if (request()->user('admin')->cannot('delete', Categoria::class)) { abort(403); }
+
         $categoria->delete();
 
         Registro::registrar_accion($categoria, 'Elimina una categoría');
